Correctly display changing RSSI values
Previously, a new device was added to the list if its RSSI value
changed. This commit fixes the issue by using the device's identifier
as key to access the map of devices. Furthermore, the results are now
sorted so that the device with the greatest signal strength is
displayed first in the list.